Create app.rb file to separate logic from input and request code
